Periodic clock source based on system time
Create a Clock trait with the methods I think a sequencer will need.
Write a concrete SystemClock struct that implements the Clock trait.
SystemClock will generate tick events by spawning a thread and polling
the OS system time to see if a set period has elapsed. It will then
execute a closure or callback function and update some internal state
variables.

Note: this approach will saturate a CPU core because the thread will
never yield. This is inefficient, but more accurate than relying on the
kernel to wake a sleeping thread.

With this callback based approach, it is critical that the work done by
the closure or callback finishes before the next tick. Otherwise time
will be lost.

Use the builder pattern for creating and configuring the clock. This
allows for one-line configuration, or updating parameters after
creation. Note: the clock must be stopped and restarted for changes to
take effect.