Create Sequencer struct and implementation
User's will interact with this real time step sequencer through the
Sequencer struct. This struct has four tracks of 16 steps. Each step con
tains a vector of Note objects or Param objects. Processing each step
generates a vector of some number of Events. An Event can be a NoteOn,
NoteOf, or ControllerChange. Bounded integers of 2, 4, or 7 bits are used
to constrain values to appropriate ranges.

Each track is polyphonic, in the sense that there is no limit to how
many Notes or Params each step can contain. However, each Note per step
must have a unique pitch so that duplicate NoteOn events to not clobber
NoteOff events. Similarly, each Param must have a unique Controller type
so that Params in the same step do not conflict with eachother.

Adding a Note or Param to a step will overwrite an existing Note or
Param with the same pitch or Controller value. This happens silently and
does not produce an Error.