Add tests for Sequencer
Provide test coverage for the Sequencer API. Test adding and remove
notes and params, receiving correct Event output, pausing and resuming
playback, and timing.

Note: The first step happens immediatley on start() then waits for a
period to pass. This means in a span of two periods, there will be three
step events. In the tests, sleep 2.5 periods to capture 3 step events.